Remove any managed references when a handle is disposed
Added an event to the HandleSuite class to inform managed users
when a handle allocated by the suite is disposed.

This allows the descriptor suites to remove an entry in the
descriptorHandles dictionary when the underlying handle is disposed and
fixes a rare crash if the OS recycles a freed handle value that was not
removed from the descriptorHandles dictionary.

Implemented IDisposable on the descriptor suites to unsubscribe from
the SuiteHandleDisposed event.
Updated the LoadPsFilter class to dispose of the descriptor suites.