Move the SPBasic suite into a separate file
Moved the SPBasic suite callbacks out of the LoadPsFilter class into a
SPBasicSuiteProvider class.
Added an IPICASuiteDataProvider interface to allow the SPBasic suite to
access the data it requires in the LoadPsFilter class.
Added an IPropertySuite interface to allow the SPBasic suite to create
the PropertyProcs structure.

Removed the PICASuites class because the SPBasicSuiteProvider handles
the PICA suite creation directly.
Changed the PICAHandleSuite to a sealed class and added a handleProcs
field to provide access to the HandleSuite pointers.
Updated the PICAUIHooksSuite to use the IPICASuiteDataProvider
interface.

Updated the PropertySuite class to implement the IPropertySuite
interface.
Updated the LoadPsFilter class to implement the IPICASuiteDataProvider
interface and use the SPBasicSuiteProvider class.