Add DPI information to the SurfaceBase class
Added DpiX and DpiY properties to the SurfaceBase class.
Added a new constructor overload to set the DPI information in the
SurfaceBGRA32, SurfaceBGRA64, SurfaceGray16 and SurfaceGray8 classes.

Updated the SurfaceFactory CreateFromGdipBitmap, CreateFromBitmapSource
methods to use the source image DPI information.
Added a CreateFromImageMode overload that sets the DPI information and
updated the LoadPsFilter constructor to use the new overload.