Add ToGdipBitmap and ToBitmapSource methods
Replaced the CreateAliasedBitmap and CreateAliasedBitmapSource methods
in the SurfaceBase class with ToGdipBitmap and ToBitmapSource methods.
Updated the SurfaceBGRA32, SurfaceBGRA64, SurfaceGray16 and
SurfaceGray8 classes to implement the new methods.

The CreateAliasedBitmap method in the SurfaceBGRA32 class has not been
removed because the DisplayPixelsProc method in LoadPsFilter uses it to
write the filter preview image to the output HDC.

The ToGdipBitmap and ToBitmapSource methods create a new image and copy
the existing pixels into it instead of reusing the memory allocated by
the existing surface.
This reduces memory usage when the destination image does not have
transparency and allows the PSFilterHost class to use the created
Bitmap or BitmapSource without having to clone it.