Improve the EnableInfo condition parsing
Added an interpreter for the enable information conditions.
This allows more conditions to be considered when deciding if a plugin
should be enabled for the current image.
For example, a plugin may specify the image should meet a minimum size
or that the host must have an active selection.

Added SupportsHostState methods the PluginData class and marked the
SupportsPixelFormat method as obsolete.