Do not try the shortcut path fix-up for 64-bit processes
The fix only applies to installers that place both the 64-bit and X86
plug-ins into a subdirectory of the 64-bit Program Files.

When loading a shortcut to these X86 plugins the WoW64 subsystem will
redirect the shortcut from the 64-bit Program Files to a nonexistent
folder in Program Files (x86).